+++ exonum/src/blockchain/service.rs
@@ -173,7 +173,7 @@ pub trait Service: Send + Sync + 'static {
     }
 
     /// A service execution. This method is invoked for each service after execution
-    /// of all transactions in the mempool but before `handle_commit` handler.
+    /// of all transactions in the block but before `handle_commit` handler.
     fn execute(&self, fork: &mut Fork) {}
 
     /// Handles block commit. This handler is invoked for each service after commit of the block.